Moving get_routing_stack() to a centralized location to avoid code duplication
Various areas of sonic-utilities are now demanding this functionality so i'm simply moving this routing to a centralized location. After spending some time debating about the ideal location for this function, we thought about sonic-config-engine/sonic_platform.py as the closest match.

Functional tests' output will be provided as part of the equivalent PR to be submitted shortly within sonic-utilities repo.

RB=
G=lnos-reviewers
R=ntrianta,pmao,rmolina,sfardeen,zxu
A=