Fixing i2c-bus-id mismtach issue with fancontrol
During the initialization of platform-monitor container (pmon), a platform-specific file utilized by 'fancontrol' is copied into pmon's /etc/ path. This file defines various variables related to fan-management aspects such as temperature thesholds and associated power outputs. This config file also specify the device in charge of performing temperature readings in the system, as well as how/where this one is connected within the i2c-buses.

The i2c-bus-id value predefined in fancontrol config-file may not match the one advertised by the i2c-driver. In this case 'fancontrol' process will not be able to execute, and in consequence the fans in the system will spin at a suboptimal speed (max speed).

This patch is simply adjusting /etc/fancontrol file in case an i2c-busid mismatch is detected.

RB=1374020
G=lnos-reviewers
R=pchaudhary,pmao,rmolina,samaity,sfardeen,zxu
A=