[FRR] Fix the traffic drop issue when bgp docker warm-restarted
This is due to the restarting router withdrawn zebra learnt routes from peers.

On FRR <= 7.3, Full GR and "Route Selection Deferral Timer" are not implemented.
So we levearge "update-delay" for restarting router to be in read-only mode
after being restarted, then checking the EoRs (implicit and explicit) from
peers or wait for the update-delay timer is expired.
There was a timing issue where the first condition above is met but zebra hasn't
learnt the local routes (e,g. connected routes), this will introduce issues
if we send updates to peers at the moment, it will withdraw the those routes
from peers thus dropping traffic.
This is a workaround to only wait for update-delay time-out to get out of the
read-only mode. This will introduce a slightly delay, but since the DUT is
restarted and the timer is small in LinkedIn DC cases (e,g, 10 seconds),
the impact is minimal if at all.

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>

RB=2243061
G=lnos-reviewers
R=pchaudha,pmao,vapatil,zxu
A=