redis-dump-load: Pin the redis package to use 3.5.3 (#9001)
Redis 4.0.0b1 has been uploaded to pip as a prerelease version. This
version drops support for Python 2 and only supports Python 3. Because
setup.py is being run, it will use the latest version of a package and
not the latest stable version (which is still 3.5.3).

Therefore, pin the redis package to version 3.5.3, so that it will work
for both Python 2 and 3.

Make sure that redis-dump-load for Python 2 builds today.

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>

RB=2942507
G=lnos-reviewers
R=pchaudha,pmao,samaity,stexu,zxu
A=