[frr/patch]: Tear down ibgp session immediately when link goes down
The standard expectation is to rely on next-hop resolution to withdraw unreachable
routes from iBGP neighbor when underlying link goes down, but this behavior is not
inline with our network design, causing extended traffic loss as a result.
To address the problem, we will introduce a LinkedIn-specific patch to change the
behavior, i.e., tear down the iBGP session immediately in response to underlying
link down notification.
The change from the patch is relatively straightforward and, more importantly, safe,
considering:
 - it only affects the link down event for iBGP sessions
 - the iBGP sessions we have between CSWs and ASWs are all over direct links
 - we have embraced this behavior since onboarding SONiC thanks to the bug prior to 3.x

Signed-off-by: Stephen Xu <stexu@linkedin.com>

RB=3245423
G=lnos-reviewers
R=pchaudhary,pmao,samaity,zxu
A=