[Seastone/3232C] Set the NUM_ECMP_MEMBERS to 64 for backwards compatibility
In SAI 3.5 (SONiC 2.x) by default we are supporting 256 ECMP groups
with 64 members each. However in SAI 4.3 (SONiC 3.x) default behavior
got changes to 128 groups and 128 members each.
This change is to make sure we are using same ECMP group/member
for SAI 4.3 (SONiC 3.x) so that the behavior is consistent.

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>

RB=3311143
G=lnos-reviewers
R=pchaudha,pmao,samaity,stexu,zxu
A=