[sysctl-net.conf]: update net.core.rmem_max setting for increased SO_RCVBUF size
The netlink SO_RCVBUF size in lldpd is currently set to 8MB. To achieve
that, the value in /proc/sys/net/core/rmem_max file must be adjusted as
well to 4MB (before kernel doubles it). This change updates the maximum
of both send and receive sizes.

Signed-off-by: Stephen Xu <stexu@linkedin.com>

RB=3588646
G=lnos-reviewers
R=pchaudhary,pmao,samaity,zxu
A=